pyproject.toml
[project]
name = "youtube-streaming-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "tweepy>=4.15.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
]
